Extract booking and gathering methods to ActionUtils
